+++ opal/static/js/test/episode_visibility.service.test.js
@@ -17,7 +17,6 @@ describe('episodeVisibility', function(){
 
       $scope = {
         currentTag: 'micro',
-        currentSubTag: 'all',
         query: {
             hospital_number: '',
             ward: ''