+++ changelog.md
@@ -1,5 +1,8 @@
 ### 0.18.4 (Minor Release)
 
+#### Removes scroll to top
+Removes the directory `scrolllto-top` as this is not commonly used. `go-to-top` is directive that is a a drop in replacement that will take the user up to the top of the page but without scrolling.
+
 #### Remove shortDate and shortDateTime
 The angular filters shortDate and shortDateTime have been deprecated for some time.